sti correction

+++ includes/corewar.h
@@ -6,7 +6,7 @@
 /*   By: sgauguet <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/12 16:53:01 by sgauguet          #+#    #+#             */
-/*   Updated: 2018/07/19 16:20:22 by sgauguet         ###   ########.fr       */
+/*   Updated: 2018/07/19 16:35:02 by sgauguet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ typedef struct	s_process
 typedef struct	s_stack
 {
 	t_process		*first_process;
-	int				nb_process;
+	unsigned long	nb_process;
 	unsigned long	process_id;
 }				t_stack;
